// src/Frontend/src/interface.ts
export interface SupplierListQuery {
    id: number;
    name: string;
    address: string;
    email: string;
    phone: string;
  }

 export interface EmployeeListQuery {
    id: number;
    code: string;
    firstName: string;
    lastName: string;
    address: string;
    email: string;
    phone: string;
  }